refactor: move & rename date-to-string util to reduce inconsistency (#22)
